Create cluster store
