+++ src/lib/stores/data-converter-config.ts
@@ -1,7 +1,5 @@
-import { browser } from '$app/env';
 import { writable } from 'svelte/store';
-
-import type { Writable } from 'svelte/store';
+import { persistStore } from '$lib/utilities/persist-store';
 
 export const dataConverterPort = persistStore('port', null);
 
@@ -15,31 +13,4 @@ export function setLastDataConverterFailure(): void {
 
 export function setLastDataConverterSuccess(): void {
   lastDataConverterStatus.set('success');
-}
-
-export function persistStore(
-  name: string,
-  initialValue: string | null = '',
-): Pick<Writable<string>, 'subscribe' | 'set'> {
-  let initialStoreValue = initialValue;
-  if (browser) {
-    try {
-      initialStoreValue = JSON.parse(window?.localStorage?.getItem(name));
-    } catch (_err) {
-      initialStoreValue = null;
-    }
-  }
-
-  const { subscribe, set } = writable(initialStoreValue);
-
-  return {
-    subscribe,
-
-    set: (x: string | null) => {
-      if (browser) {
-        window?.localStorage?.setItem(name, JSON.stringify(x));
-      }
-      set(x);
-    },
-  };
 }
